+++ src/elf.cpp
@@ -820,7 +820,7 @@ char *elfReadString(u8 *data, int *bytesRead)
     *bytesRead = 1;
     return NULL;
   }
-  *bytesRead = strlen((char *)data) + 1;
+  *bytesRead = (int)strlen((char *)data) + 1;
   return (char *)data;
 }
 
@@ -1070,7 +1070,7 @@ void elfParseCFA(u8 *top)
   ELFcie *cies = NULL;
   
   while(data < end) {
-    u32 offset = data - topOffset;
+    u32 offset = (u32)(data - topOffset);
     u32 len = elfRead4Bytes(data);
     data += 4;
 
@@ -1110,7 +1110,7 @@ void elfParseCFA(u8 *top)
       cie->returnAddress = *data++;
 
       cie->data = data;
-      cie->dataLen = dataEnd - data;
+      cie->dataLen = (u32)(dataEnd - data);
     } else {
       ELFfde *fde = (ELFfde *)calloc(1, sizeof(ELFfde));
 
@@ -1136,7 +1136,7 @@ void elfParseCFA(u8 *top)
       data += 4;
 
       fde->data = data;
-      fde->dataLen = dataEnd - data;
+      fde->dataLen = (u32)(dataEnd - data);
 
       if((elfFdeCount %10) == 0) {
         elfFdes = (ELFfde **)realloc(elfFdes, (elfFdeCount+10) *
@@ -2762,7 +2762,7 @@ bool elfReadProgram(ELFHeader *eh, u8 *data, int& size, bool parseDebug)
     
     while(ddata < end) {
       unit = elfParseCompUnit(ddata, abbrevdata);
-      unit->offset = ddata-debugdata;
+      unit->offset = (u32)(ddata-debugdata);
       elfParseLineInfo(unit, data);
       if(last == NULL)
         elfCompileUnits = unit;
